APPS CON ESTRUCTURA/GRAFOLEX-INSIGHT/requirements.txt
gradio
torch
transformers
opencv-python
numpy
pillow
